requirements.txt
google-adk>=1.0.0
google-cloud-bigquery>=3.25.0
google-auth>=2.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.27.0
